SidebarParams: showing different options for different modes
