+++ src/components/OrderForm/style.scss
@@ -317,6 +317,7 @@
 
   &.disabled {
     opacity: 0.5;
+    cursor: not-allowed;
   }
 
   &.invalid {
