+++ src/modals/Strategy/ExecutionOptionsModal/ExecutionOptionsModal.js
@@ -1,18 +1,27 @@
 import React, { useState } from 'react'
+import { useSelector } from 'react-redux'
 import PropTypes from 'prop-types'
 import { useTranslation } from 'react-i18next'
 import Modal from '../../../ui/Modal'
 import AmountInput from '../../../components/OrderForm/FieldComponents/input.amount'
 
 import ExecutionOptionsBody from './ExecutionOptionsBody'
+import { getIsPaperTrading } from '../../../redux/selectors/ui'
+
 import './style.scss'
 
 const ExecutionOptionsModal = (props) => {
   const {
-    isOpen, onClose, capitalAllocation, setCapitalAllocation, stopLossPerc,
-    maxDrawdownPerc, startExecution,
+    isOpen,
+    onClose,
+    capitalAllocation,
+    setCapitalAllocation,
+    stopLossPerc,
+    maxDrawdownPerc,
+    startExecution,
   } = props
   const [capitalAllocationError, setCapitalAllocationError] = useState('')
+  const isPaperTrading = useSelector(getIsPaperTrading)
 
   const { t } = useTranslation()
 
@@ -41,17 +50,29 @@ const ExecutionOptionsModal = (props) => {
     <Modal
       isOpen={isOpen}
       onClose={onClose}
-      label={t('strategyEditor.executionOptionsModal.title')}
+      label={
+        isPaperTrading
+          ? t('strategyEditor.executionOptionsModal.title')
+          : t('strategyEditor.executionOptionsModal.disabledTitle')
+      }
       onSubmit={onSubmit}
     >
-      <ExecutionOptionsBody capitalAllocationHandler={capitalAllocationHandler} capitalAllocationError={capitalAllocationError} t={t} {...props} />
+      <ExecutionOptionsBody
+        isPaperTrading={isPaperTrading}
+        capitalAllocationHandler={capitalAllocationHandler}
+        capitalAllocationError={capitalAllocationError}
+        t={t}
+        {...props}
+      />
       <Modal.Footer>
         <Modal.Button secondary onClick={onClose}>
           {t('ui.closeBtn')}
         </Modal.Button>
-        <Modal.Button primary onClick={onSubmit} disabled={!isFullFilled}>
-          {t('ui.startBtn')}
-        </Modal.Button>
+        {isPaperTrading && (
+          <Modal.Button primary onClick={onSubmit} disabled={!isFullFilled}>
+            {t('ui.startBtn')}
+          </Modal.Button>
+        )}
       </Modal.Footer>
     </Modal>
   )
