+++ src/ui/Input/style.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import "../../variables.scss";
 
 .hfui-input {
   p {
@@ -17,6 +17,7 @@
     line-height: 12px;
     outline: none;
     border-radius: $border-radius;
+    cursor: inherit;
   }
 }
